[#18] rename : 패키지명 수정 및 파일 위치 이동
