+++ src/main/kotlin/com/flabedu/blackpostoffice/commom/annotation/LoginCheck.kt
@@ -1,4 +1,4 @@
-package com.flabedu.blackpostoffice.util.annotation
+package com.flabedu.blackpostoffice.commom.annotation
 
 import com.flabedu.blackpostoffice.domain.model.User.Role
 import com.flabedu.blackpostoffice.domain.model.User.Role.NONE
