+++ src/main/kotlin/com/flabedu/blackpostoffice/commom/encryption/Encryption.kt
@@ -1,4 +1,4 @@
-package com.flabedu.blackpostoffice.util.encryption
+package com.flabedu.blackpostoffice.commom.encryption
 
 interface Encryption {
