+++ src/main/kotlin/com/flabedu/blackpostoffice/commom/logging/GlobalLogger.kt
@@ -1,4 +1,4 @@
-package com.flabedu.blackpostoffice.util.logging
+package com.flabedu.blackpostoffice.commom.logging
 
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
