+++ src/main/kotlin/com/flabedu/blackpostoffice/commom/utils/constants/LoginContants.kt
@@ -1,4 +1,4 @@
-package com.flabedu.blackpostoffice.util.interceptor
+package com.flabedu.blackpostoffice.commom.utils.constants
 
 const val LOGIN_MY_EMAIL = "email"
 const val MY_ROLE = "role"
