+++ src/main/kotlin/com/flabedu/blackpostoffice/service/UserService.kt
@@ -3,7 +3,7 @@ package com.flabedu.blackpostoffice.service
 import com.flabedu.blackpostoffice.controller.dto.UserDto
 import com.flabedu.blackpostoffice.domain.mapper.UserMapper
 import com.flabedu.blackpostoffice.exception.user.DuplicateEmailException
-import com.flabedu.blackpostoffice.util.encryption.Sha256Encryption
+import com.flabedu.blackpostoffice.commom.encryption.Sha256Encryption
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
